+++ paws.common/R/handlers_core.R
@@ -2,17 +2,6 @@
 
 #' @include util.R
 
-# Add the name/version to the User-Agent request header.
-sdk_version_user_agent_handler <- function(request) {
-  paws_version <- .__NAMESPACE__.[["spec"]]["version"] %||% packageVersion("paws.common")
-  user_agent <- sprintf(
-    "paws/%s (R%s; %s; %s)",
-    paws_version, getRversion(), R.version$os, R.version$arch
-  )
-  request$http_request$header["User-Agent"] <- user_agent
-  return(request)
-}
-
 # Add the execution environment to the User-Agent request header.
 # TODO: Implement.
 add_host_exec_env_user_agent_handler <- function(request) {