+++ paws.common/R/util.R
@@ -242,3 +242,12 @@ parse_in_half <- function(x, char = "=") {
     substr(x, right_start, right_end)
   )
 }
+
+set_user_agent <- function(pkgname) {
+  paws_version <- .__NAMESPACE__.[["spec"]]["version"] %||% packageVersion(pkgname)
+  user_agent <- sprintf(
+    "paws/%s (R%s; %s; %s)",
+    paws_version, getRversion(), R.version$os, R.version$arch
+  )
+  assign("PAWS_USER_AGENT", user_agent, envir = getNamespace(pkgname))
+}